Add delete workflow (#3)
This workflow allows deleting VMs from a group of hosts.
A user specifies the number of VMs to delete and the playbook ensures that this number is deleted.
If more VMs are requested to be deleted than exist, the playbook will fail.

A future improvement would be to allow deleting all VMs from a group without specifying a number.